fix(grading): allow a manual override for a student with no grades yet
report_scores rows are only created by recomputeReportScores, for
students who already have at least one component grade. Setting a
manual override for a student with none (a transfer student) called
SetManualReportScore, an UPDATE, which found no row and surfaced as
pgx.ErrNoRows -> 404 ASSESSMENT_COMPONENT_NOT_FOUND -- both wrong and
misleading.

SetManualScore now falls back to inserting the row, with the
automatic value the student would otherwise have (studentAutomaticScore,
reusing recomputeReportScores' rule), then applying the override via
the existing UpsertReportScore query. Clearing the override afterwards
goes through the normal UPDATE path against the now-existing row and
restores that same automatic baseline.

While exercising UpsertReportScore against a real database, its SQL
turned out to have a latent, unrelated bug: `coalesce($8, $9)` in the
insert's final_score expression left Postgres unable to deduce a type
for $8, failing every call (including the pre-existing recompute path,
apparently never hit by a real database before). Fixed with explicit
::numeric casts and regenerated via sqlc.

Added a real-database integration test (grading_integration_test.go,
following the attendance module's testcontainers pattern, since
grading had no service-level test scaffold yet) covering both setting
the override on a gradeless student and clearing it back to automatic.

+++ apps/api/internal/modules/grading/queries/reports.sql
@@ -5,7 +5,7 @@
 -- automatic_score. This is what lets SetManualReportScore restore the
 -- automatic value by clearing the override, without a full recompute.
 insert into report_scores (tenant_id, academic_year_id, term_id, class_id, subject_id, student_user_id, previous_score, manual_score, automatic_score, final_score, computed_at)
-values ($1, $2, $3, $4, $5, $6, $7, $8, $9, coalesce($8, $9), now())
+values ($1, $2, $3, $4, $5, $6, $7, $8, $9, coalesce($8::numeric, $9::numeric), now())
 on conflict (academic_year_id, term_id, class_id, subject_id, student_user_id) do update
   set previous_score = excluded.previous_score,
       manual_score = coalesce(excluded.manual_score, report_scores.manual_score),
